% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fault.R
\name{calib}
\alias{calib}
\title{Calibrate the algorithm}
\usage{
calib(database, dif_per_1, dif_per_2, dif_per_3)
}
\arguments{
\item{database}{csv file with the profiles}

\item{dif_per_1}{lower percentile to remove in the difference stage}

\item{dif_per_2}{upper percentile to remove in the difference stage}

\item{dif_per_3}{percentile used in fine tuning (should be higher than per_2)}
}
\description{
Calibrate the algorithm
}
