% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fault.R
\name{denoise}
\alias{denoise}
\title{Calibrate the algorithm}
\usage{
denoise(database, dt_1, dt_2, dt_3, t1, t3)
}
\arguments{
\item{database}{csv file with the profiles}

\item{dt_1}{lower differential threshold}

\item{dt_2}{upper differential threshold}

\item{dt_3}{fine tuned differential threshold}

\item{t1}{upper boxplot threshold}

\item{t3}{lower boxplot threshold}
}
\description{
Calibrate the algorithm
}
